Moved marked from custom to internal
