+++ src/com/content-comments/comments-markup.js
@@ -5,7 +5,7 @@ import NavLink							from 'com/nav-link/link';
 import SVGIcon							from 'com/svg-icon/icon';
 import InputTextArea					from 'com/input-textarea/input-textarea';
 
-import marked 								from '../../custom/marked/marked';
+import marked 								from '../../internal/marked/marked';
 
 export default class ContentCommentsMarkup extends Component {
 	constructor( props ) {
