+++ src/com/content-event/event-home.js
@@ -5,7 +5,7 @@ import NavLink 							from 'com/nav-link/link';
 import ButtonBase						from 'com/button-base/base';
 
 import $Theme							from '../../shrub/js/theme/theme';
-import marked 								from '../../custom/marked/marked';
+import marked 								from '../../internal/marked/marked';
 
 
 export default class ContentEventHome extends Component {
