+++ src/internal/marked/InlineLexer.js
