+++ src/internal/marked/Lexer.js
