+++ src/internal/marked/Parser.js
