+++ src/internal/marked/marked.js
